+++ Students/Churkin Anton/project/webpack/src/public/index.html
@@ -1,0 +1,21 @@
+<!DOCTYPE html>
+<html>
+
+<head>
+    <title>eShop</title>
+</head>
+
+<body>
+    <h1>eShop</h1>
+    <div class="goods"></div>
+
+    <hr>
+    <div>
+        <button class="cart-button" type="button">Корзина</button>
+    </div>
+
+    <div id="cart" class="invisible">
+    </div>
+</body>
+
+</html>
