--- Students/Churkin Anton/project/webpack/style/style.css
@@ -1,3 +1,0 @@
-.invisible{
-    display: none;
-}
